budidaya_cabe_streamlit/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
prophet>=1.1.0
scikit-learn>=1.3.0
requests>=2.31.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
qrcode>=7.4.2
